refactor : RecipeScreen move to AccountScreen
Composable for Liked Recipe move to AccountScreen to follow figma design
Modified test to reflect the change
